portfolio_app_v6/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
yfinance==0.2.43
python-dotenv==1.0.
